requirements.txt
flask
requests
PyJWT
pytest